feat : 코드리뷰반영, commoninput내제화, 불요코드삭제
