+++ apps/modern-tmi-client/src/app/app.tsx
@@ -1,5 +1,4 @@
 import styled from '@emotion/styled';
-import NxWelcome from './nx-welcome';
 import RegisterPage from './pages/RegisterPage';
 
 const StyledApp = styled.div`
